requirements.txt
h5py==3.13.0
matplotlib==3.10.3
numpy==2.2.6
PyYAML==6.0.2
PyYAML==6.0.2
sympy==1.13.3
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
